Conflicts:
	cmake/GenerateGoogleTest.cmake

+++ src/ops.h
@@ -32,7 +32,7 @@ nd_t ndfmad       (nd_t z,float a,const nd_t x,float b,const nd_t y,size_t ndim,
 nd_t ndxor_ip     (nd_t z,uint64_t c,size_t ndim, size_t* shape);
 nd_t ndconvert_ip (nd_t z, nd_type_id_t type);
 
-nd_t ndaffine (nd_t dst, const nd_t src, const double *transform, const nd_affine_params_t *params);
+nd_t ndaffine     (nd_t dst, const nd_t src, const float *transform, const nd_affine_params_t *params);
 #ifdef __cplusplus
 } //extern "C" {
 #endif
